requirements.txt
# Core Pydantic AI and async support
pydantic-ai[openai,gemini]>=0.0.14
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP client and web requests
httpx>=0.25.0
aiohttp>=3.8.0

# Search APIs
tavily-python>=0.3.0
google-search-results>=2.4.2  # SerpAPI

# Gmail API integration
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0

# LLM Providers
openai>=1.0.0
google-generativeai>=0.7.0

# CLI and utilities
click>=8.0.0
rich>=13.0.0
python-dotenv>=1.0.0

# Testing and development
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Code quality
ruff>=0.1.0
mypy>=1.5.0
black>=23.0.0

# Logging and monitoring
structlog>=23.0.0